Remove password.reset.token.reset
This was only documented as a concession to descriptiveness over
perscriptiveness, in the face of me wanting to get every field
documented when implementing the fields-must-be-documented test,
without having to take time to rethink the structure of any
existing, undocumented profile data (I was out to dinner, we'd
just paid the check, and we had a bus to catch). It was only ever
added to describe Wikimedia (or MediaWiki)'s password reset
behavior, and a properly-demarcated password.reset.token.login value
communicates that effect just as accurately.